requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-openai
faiss-cpu
huggingface_hub
sentence-transformers
openai
python-dotenv
pypdf
python-docx
docx2txt
pandas
openpyxl
markdown
Pillow
requests